requirements.txt
Django==2.2.7
Pillow==6.2.1
django-redis==4.11.0
django-registration==3.0.1
django-storages==1.8.0
huey==2.1.3
monero==0.6.2
psycopg2==2.8.4
pysha3==1.0.2
redis==3.3.11
